# srcs/api-gateway/openapi.yaml
...
openapi: 3.1.0
info:
  title: CRUD Master API Gateway
  description: |
    API Gateway for the CRUD Master microservices infrastructure.
    This gateway routes requests to the Inventory API (HTTP) and Billing API (RabbitMQ).
    
    ## Architecture
    - **Inventory API**: Handles movie inventory CRUD operations via HTTP proxy
    - **Billing API**: Processes payment orders via RabbitMQ message queue
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: http://gateway-vm:3000
    description: Gateway VM server
...
paths:
...
  /api/billing:
    post:
      tags:
        - Billing
      summary: Create a new order
      description: |
        Send an order to the billing queue via RabbitMQ. The order will be processed asynchronously.
        The gateway will return success immediately, even if the billing service is not running.
        When the billing service starts, it will process all queued messages.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
            example:
              user_id: "3"
              number_of_items: "5"
              total_amount: "180"
      responses:
        '200':
          description: Order sent to billing queue successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Order sent to billing queue."
        '500':
          description: Failed to send message to RabbitMQ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Connection refused"
components:
  schemas:
...
    OrderInput:
      type: object
      properties:
        user_id:
          type: string
          description: ID of the user making the order
          example: "3"
        number_of_items:
          type: string
          description: Number of items in the order
          example: "5"
        total_amount:
          type: string
          description: Total cost of the order
          example: "180"
      required:
        - user_id
        - number_of_items
        - total_amount
...